requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
astunparse==1.6.3
attrs==23.2.0
blis==0.7.11
cachetools==5.4.0
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
datasets==2.20.0
dill==0.3.8
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl#sha256=6a0f857a2b4d219c6fa17d455f82430b365bf53171a2d919b9376e5dc9be032e
et-xmlfile==1.1.0
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.5.0
gast==0.6.0
gensim==4.3.3
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
google-pasta==0.2.0
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.62.2
h5py==3.11.0
httplib2==0.22.0
huggingface-hub==0.24.0
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
joblib==1.4.2
keras==3.4.1
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
libclang==18.1.1
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.4.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
namex==0.0.8
networkx==3.3
nltk==3.8.1
numpy==1.26.4
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
plotly==5.23.0
preshed==3.0.9
proto-plus==1.24.0
protobuf==4.25.3
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pydantic==2.8.2
pydantic_core==2.20.1
pygame==2.6.0
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-louvain==0.16
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.13.1
sentence-transformers==3.0.1
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.4
srsly==2.4.8
sympy==1.13.0
tbb==2021.13.0
tenacity==8.5.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
tf_keras==2.17.0
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.4
transformers==4.42.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
